Fix /etc/hosts networking documentation
Clarifies that adding `/etc/hosts` entries for `kafka` and `minio` can break
container-to-container communication. Updates README.md files
to indicate that `/etc/hosts` modifications are optional, not required.
The previous instructions could cause `kafka-create-topics` to fail with
connection errors when trying to connect to the Kafka broker.

Error resolved:
`Connection to node -1 (kafka/127.0.0.1:29092) could not be established. Broker may not be available.`

Signed-off-by: Gloria Ciavarrini <gciavarrini@redhat.com>